bot/requirements.txt
python-dotenv==0.20.0
aiohttp==3.8.1
aiogram==2.21
